Improve the usage example in README.md
Most importantly: The graph never drew the inbytes line, because the
call to line() was missing. Also show how the RrdGraphDef datasource
is coupled with the line (via "inbytes-average" in the example now).

Make it clear how the RrdDb and RrdGraphDef are coupled: Via the
rrdPath.

Don't call setTime() because 1) 't' is nowhere defined nor declared in
the example, thus the reader has now idea of its type and 2) because
the default behavior of the sample constructor, setting the time to
'now', is fine for most use cases.

Similar, don't call setStartTime() and setEndTime() as the default
RrdGraphDef() behavior, creating a graph for the last 24h, is just
fine for an example

Also invoke setFilename with something sensible.