Add OSGi metadata to the main rrd4j jar
In order to make deployment in OSGi enviroments simple specific OSGi
metadata needs to be added to the Jar file. Instead of maintaining
these by hand, a Maven plugin is typically used.

This commit introduces the usage of the bnd-maven-plugin, with the
following notes:

- MongoDB, Sleepycat and NIO backends are optional since not
  all installations have them available. These are marked as
  optional imports to allow the bundle to resolve even when
  they are not present
- All packages are exported except for demo, converter and inspector,
  following the way the main jar is built

I have tested the usage example present in the README file and
things work as expected.