Minor rendering improvements
This is a collection of minor changes we use in I2P.
Some of it is a matter of taste or may be different than what jrobin did.
Other changes omit output when the alpha or stroke is 0,
which is useful for SVG output but is more efficient regardless.
Please review and cherrypick what you like and skip what you don't.

- Change u to micron in legend to match value axis
- Combine drawRules() and drawSpans() to only iterate through plotElements once
- Only clip and reset in drawRulesAndSpans() if we output something
- Move title up a little for better padding below
- Position long titles to clip only on right side
- Don't draw arrowheads if disabled by color alpha of 0
- Don't draw ticks if disabled by stroke width of 0
- Don't draw three legend color boxes on top of each other unless all have some transparency
- Replace NaN with 0 in legend values
- Output milli values as 0.xxx rather than xxx m
- Add day number to time axis day names
- Javadoc warning fixes